temp dir creation at first run

+++ index.js
@@ -13,8 +13,15 @@ module.exports = {
 
 if (!fs.existsSync(data_dir)){
     fs.mkdirSync(data_dir);
-    fs.writeFileSync(data_dir + '/prefixes.json', "{}", { flag: "wx" });
-    fs.writeFileSync(data_dir + '/autoresponder.json', "{}", { flag: "wx" });
+
+    if (!fs.existsSync(data_dir+"/prefixes.json"))
+        fs.writeFileSync(data_dir + '/prefixes.json', "{}", { flag: "wx" });
+
+    if (!fs.existsSync(data_dir+"/autoresponder.json"))
+        fs.writeFileSync(data_dir + '/autoresponder.json', "{}", { flag: "wx" });
+
+    if (!fs.existsSync(data_dir+"/temp"))
+        fs.mkdirSync(data_dir+"/temp");
 }
 
 // end
